refactored httpserver examples
