+++ examples/web/httpserver/hello/static/hello.txt
