+++ examples/web/httpserver/hello/static/index.html
