+++ examples/web/httpserver/statik/README.md
