+++ examples/web/httpserver/statik/main_test.go
