+++ examples/web/httpserver/statik/static/index.html
